requirements.txt
discord.py==1.0.1
redis==3.0.1
python-twitch-client==0.6.0
